Coursework: added stripe.

+++ Coursework/domain/src/main/java/org/geekhub/ticketbooking/show/Show.java
@@ -2,7 +2,6 @@ package org.geekhub.ticketbooking.show;
 
 import org.geekhub.ticketbooking.movie.Movie;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
@@ -11,7 +10,7 @@ import java.util.Objects;
 
 public class Show {
     private int id;
-    private BigDecimal price;
+    private Long price;
     private OffsetDateTime start;
     private OffsetDateTime end;
     private Movie movie;
@@ -19,14 +18,14 @@ public class Show {
 
     public Show() {
         this.id = -1;
-        this.price = BigDecimal.valueOf(0);
+        this.price = 0L;
         this.start = null;
         this.end = null;
         this.movie = new Movie();
         this.hallId = -1;
     }
 
-    public Show(int id, BigDecimal price, OffsetDateTime start, OffsetDateTime end,
+    public Show(int id, Long price, OffsetDateTime start, OffsetDateTime end,
                 Movie movie, int hallId) {
         this.id = id;
         this.price = price;
@@ -44,11 +43,11 @@ public class Show {
         this.id = id;
     }
 
-    public BigDecimal getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
